+++ internal/component/common/loki/client/endpoint.go
@@ -11,7 +11,7 @@ import (
 	"github.com/grafana/dskit/backoff"
 
 	"github.com/grafana/alloy/internal/component/common/loki"
-	"github.com/grafana/alloy/internal/component/common/loki/client/internal"
+	"github.com/grafana/alloy/internal/component/common/loki/client/internal/marker"
 )
 
 type endpoint struct {
@@ -27,7 +27,7 @@ type endpoint struct {
 	backoff *backoff.Backoff
 }
 
-func newEndpoint(metrics *metrics, cfg Config, logger log.Logger, markerHandler internal.MarkerHandler) (*endpoint, error) {
+func newEndpoint(metrics *metrics, cfg Config, logger log.Logger, markerHandler marker.MarkerHandler) (*endpoint, error) {
 	logger = log.With(logger, "component", "endpoint", "host", cfg.URL.Host)
 
 	shards, err := newShards(metrics, logger, markerHandler, cfg)
