+++ internal/component/common/loki/client/internal/marker/marker_encoding.go
@@ -1,4 +1,4 @@
-package internal
+package marker
 
 import (
 	"encoding/binary"
