+++ internal/component/common/loki/client/internal/marker/marker_encoding_test.go
@@ -1,4 +1,4 @@
-package internal
+package marker
 
 import (
 	"testing"
