+++ internal/component/common/loki/client/internal/marker/marker_file_handler.go
@@ -1,4 +1,4 @@
-package internal
+package marker
 
 import (
 	"bytes"
