+++ internal/component/common/loki/client/internal/marker/marker_handler_test.go
@@ -1,4 +1,4 @@
-package internal
+package marker
 
 import (
 	"os"
