+++ internal/component/common/loki/client/internal/marker/metrics.go
@@ -1,4 +1,4 @@
-package internal
+package marker
 
 import (
 	"github.com/grafana/alloy/internal/util"
